refactor(analysis-types): extract effort results DTO
